Slience some wxWarnings about modality settings, Closes: #11
The actual problem is that the modality settings contain references
to environment variables that are not set, and the wxXml component
warns about this at logging level wxLOG_Warning. The workaround is
to set thelog level to wxLOG_Error before reading the actual config
entry and reset it to its original value afterwards.